+++ esbuild.config.cjs
@@ -1,7 +1,9 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 const esbuild = require('esbuild');
-const plugin = require('node-stdlib-browser/helpers/esbuild/plugin');
-const stdLibBrowser = require('node-stdlib-browser');
+// Bring back if needed
+// const plugin = require('node-stdlib-browser/helpers/esbuild/plugin');
+// const stdLibBrowser = require('node-stdlib-browser');
+const fs = require('fs');
 
 esbuild
   .build({
@@ -13,11 +15,19 @@ esbuild
       'process.env.NODE_ENV': 'production',
     },
     format: 'esm',
-    plugins: [plugin(stdLibBrowser)],
+    // Bring back if needed
+    // plugins: [plugin(stdLibBrowser)],
     entryPoints: ['./src/elven.ts'],
     bundle: true,
+    metafile: true,
     minify: true,
     outdir: 'build',
     platform: 'browser',
+  })
+  .then((result) => {
+    return esbuild.analyzeMetafile(result.metafile);
+  })
+  .then((result) => {
+    fs.writeFileSync('./build/meta.txt', result);
   })
   .catch(() => process.exit(1));