+++ src/auth/login-with-extension.ts
@@ -1,7 +1,7 @@
 import { ls } from '../utils/ls-helpers';
 import { initExtensionProvider } from './init-extension-provider';
 import { errorParse } from '../utils/error-parse';
-import { LoginMethodsEnum } from '../types';
+import { EventStoreEvents, LoginMethodsEnum } from '../types';
 import { getNewLoginExpiresTimestamp } from './expires-at';
 import { accountSync } from './account-sync';
 import { EventsStore } from '../events-store';
@@ -11,6 +11,7 @@ export const loginWithExtension = async (elven: any, token?: string) => {
 
   try {
     if (dappProvider) await dappProvider.login();
+    EventsStore.run(EventStoreEvents.onLoginPending);
   } catch (e) {
     const err = errorParse(e);
     console.warn(`Something went wrong trying to login the user: ${err}`);
@@ -40,21 +41,19 @@ export const loginWithExtension = async (elven: any, token?: string) => {
         throw new Error('Canceled!');
       }
 
-      EventsStore.run('onLoginPending');
-
       ls.set('address', address);
       ls.set('loginMethod', LoginMethodsEnum.browserExtension);
       ls.set('expires', getNewLoginExpiresTimestamp());
 
       await accountSync(elven);
 
-      EventsStore.run('onLoggedIn');
-
       return dappProvider;
     } catch (e: any) {
       console.warn(
         `Something went wrong trying to synchronize the user account: ${e?.message}`
       );
+    } finally {
+      EventsStore.run(EventStoreEvents.onLoggedIn);
     }
   }
 };