+++ src/types.ts
@@ -1,16 +1,19 @@
-import { ExtensionProvider } from '@multiversx/sdk-extension-provider';
+import { ExtensionProvider } from '@multiversx/sdk-extension-provider/out/extensionProvider';
 import { Transaction } from '@multiversx/sdk-core/out/transaction';
-import { WalletConnectProvider } from '@multiversx/sdk-wallet-connect-provider';
-import { WalletProvider } from '@multiversx/sdk-web-wallet-provider/out';
+import { WalletConnectV2Provider } from '@multiversx/sdk-wallet-connect-provider/out/walletConnectV2Provider';
+import { WalletProvider } from '@multiversx/sdk-web-wallet-provider/out/walletProvider';
 
 export interface InitOptions {
   apiUrl?: string;
   chainType?: string;
   apiTimeout?: number;
-  walletConnectBridgeAddresses?: string[];
+  walletConnectV2ProjectId?: string;
+  walletConnectV2RelayAddresses?: string[];
   onLoginPending?: () => void;
   onLoggedIn?: () => void;
   onLogout?: () => void;
+  onQrPending?: () => void;
+  onQrLoaded?: () => void;
   onTxStarted?: (transaction: Transaction) => void;
   onTxSent?: (transaction: Transaction) => void;
   onTxFinalized?: (transaction: Transaction) => void;
@@ -26,7 +29,7 @@ export enum LoginMethodsEnum {
 
 export type DappProvider =
   | ExtensionProvider
-  | WalletConnectProvider
+  | WalletConnectV2Provider
   | WalletProvider
   | undefined;
 
@@ -34,4 +37,20 @@ export interface LoginOptions {
   qrCodeContainer?: string | HTMLElement;
   token?: string;
   callbackRoute?: string;
+}
+
+export enum DappCoreWCV2CustomMethodsEnum {
+  erd_cancelAction = 'erd_cancelAction',
+}
+
+export enum EventStoreEvents {
+  onLoginPending = 'onLoginPending',
+  onLoggedIn = 'onLoggedIn',
+  onQrPending = 'onQrPending',
+  onQrLoaded = 'onQrLoaded',
+  onLogout = 'onLogout',
+  onTxStarted = 'onTxStarted',
+  onTxSent = 'onTxSent',
+  onTxFinalized = 'onTxFinalized',
+  onTxError = 'onTxError',
 }